Day 02 - Pseudocode started
Started working on pseudocode. Good progress.
Re-structured the Simulation class and added some ideas for generating limbs based on the parent creatures.
Jumping power of the creature is a random  number between the creature’s parents’ jumping power.
For each limb, the creature has higher probability(80%)  of generating the same limb type as the parents if they both have the same limb type, otherwise the limb generation is random.
Other creature characteristics will be calculated smiarly
All of this is not final and I will have to perform tests and check if this approach succeeds in generating successful creatures.
Other limb, wheel and leg characteristics will be simulated similarly.