Completed joint class pseudo
Joint class pseudo complete. Added mutation n the joints where if the constructor is called with parents, it has an 80% of following the parents' path.

Not sure if this is needed, maybe do it from the higher classes, so from the creature class it gives this mutation where it has an 80% of getting the parents' variables

To do:
-complete wheel class
-fix simulation class