DAY 14(22 03 2023) Classes and objects in experiment
Lots of progress
Managed to establish classes and objects in my simulkation.
Fixed the camera issue I had, I need to assign the creature BODY not the object to the variable for it to work.
Struggling with automating this in my simulation scrip tbut I think its just a simple mistake and my mind is boggled.
To do:
-	Assign the camera variable automatically
-	Figure out why numOfCreatures don’t work
-	Have it go from generation to generat
Future to-do:
-	New arena
-	New variables to consider
-	Actual experiment\