DAY17 Colors(almost) and passing values
FINALLY MANAGED TO PASS FROM CLASS TO CLASS
Noticed that mutations matter. With no mutations the creatures get stuck to only one type of creature which is basically the exact same
Also think I need to change the way the speed interacts with the creature.
WITH NO MUTATIONS THE CREATURES ENDS UP BEING DIFFERNET EACH TIME WITH THE SAME NEIRONMENT
WITH MUTAITONS IT ALWAYS ENDS UP BEING THE SAME( THE MOST EFFICIENT ONE)

Colors:
Mutaitons: shades of purple
Normal: Shades of grey
Assign colour aporpierlty? Like depending on the parents and also, the bigger the wheels, the darker the color
-	Did colors(almost
-	Values pass from gen to gen
Next:
-	Fix gen colors 100%
-	Some Interface showing generation
-	Database?

-Different tpes of wheels